+++ include/bh_python/metadata.hpp
@@ -9,17 +9,17 @@
 
 #include <pybind11/pytypes.h>
 
-inline bool PyObject_Check(void* value) { return value != nullptr; }
+struct metadata_t : py::dict {
+    PYBIND11_OBJECT(metadata_t, dict, PyDict_Check);
 
-struct metadata_t : py::object {
-    PYBIND11_OBJECT(metadata_t, object, PyObject_Check);
+    using dict::dict;
 
-    // default initialize to None
+    // default initialize to empty dict (must not be explicit)
     metadata_t()
-        : object(Py_None, borrowed_t{}) {}
+        : dict() {}
 
-    bool operator==(const metadata_t& other) const { return py::object::equal(other); }
+    bool operator==(const metadata_t& other) const { return py::dict::equal(other); }
     bool operator!=(const metadata_t& other) const {
-        return py::object::not_equal(other);
+        return py::dict::not_equal(other);
     }
 };