+++ src/boost_histogram/_internal/hist.py
@@ -1,23 +1,25 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, division, print_function
 
-from .kwargs import KWArgs
-
-from .. import _core
-from .view import _to_view
-from .axis import Axis
-from .axistuple import AxesTuple
-from .sig_tools import inject_signature
-from .storage import Double, Storage
-from .utils import cast, register, set_family, MAIN_FAMILY, set_module
-from .six import string_types
-
-import warnings
 import copy
-import numpy as np
-
 import os
 import threading
+import warnings
+
+import numpy as np
+
+from .. import _core
+from .axis import Axis
+from .axistuple import AxesTuple
+from .kwargs import KWArgs
+from .sig_tools import inject_signature
+from .six import string_types
+from .storage import Double, Storage
+from .utils import cast, register, set_family, MAIN_FAMILY, set_module
+from .view import _to_view
+
+
+NOTHING = object()
 
 _histograms = (
     _core.hist.any_double,
@@ -49,10 +51,7 @@ def _arg_shortcut(item):
     msg = "Developer shortcut: will be removed in a future version"
     if isinstance(item, tuple) and len(item) == 3:
         warnings.warn(msg, FutureWarning)
-        return _core.axis.regular_uoflow(item[0], item[1], item[2], None)
-    elif isinstance(item, tuple) and len(item) == 4:
-        warnings.warn(msg, FutureWarning)
-        return _core.axis.regular_uoflow(*item)
+        return _core.axis.regular_uoflow(item[0], item[1], item[2])
     elif isinstance(item, Axis):
         return item._ax
     else:
@@ -85,8 +84,6 @@ def _expand_ellipsis(indexes, rank):
 @set_family(MAIN_FAMILY)
 @set_module("boost_histogram")
 class Histogram(object):
-    __slots__ = ("_hist", "axes", "metadata")
-
     @inject_signature(
         "self, *axes, storage=Double(), metadata=None", locals={"Double": Double}
     )
@@ -162,12 +159,24 @@ class Histogram(object):
 
         return AxesTuple(self._axis(i) for i in range(self.ndim))
 
-    def _new_hist(self, _hist):
+    def _new_hist(self, _hist, memo=NOTHING):
         """
         Return a new histogram given a new _hist, copying metadata.
         """
 
-        other = self.__class__(_hist, metadata=self.metadata)
+        other = self.__class__(_hist)
+        for item in self.__dict__:
+            if item not in ["axes", "_hist"]:
+                if memo is NOTHING:
+                    other.__dict__[item] = self.__dict__[item]
+                else:
+                    other.__dict__[item] = copy.deepcopy(self.__dict__[item], memo)
+        other.axes = other._generate_axes_()
+        for ax in other.axes:
+            if memo is NOTHING:
+                ax._ax.metadata = copy.copy(ax._ax.metadata)
+            else:
+                ax._ax.metadata = copy.deepcopy(ax._ax.metadata, memo)
         return other
 
     @property
@@ -376,24 +385,40 @@ class Histogram(object):
         return self._new_hist(self._hist.reduce(*args))
 
     def __copy__(self):
-        other = self._new_hist(copy.copy(self._hist))
-        return other
+        return self._new_hist(copy.copy(self._hist))
 
     def __deepcopy__(self, memo):
-        other = self.__class__.__new__(self.__class__)
-        other._hist = copy.deepcopy(self._hist, memo)
-        other.metadata = copy.deepcopy(self.metadata, memo)
-        other.axes = other._generate_axes_()
-        return other
+        return self._new_hist(copy.deepcopy(self._hist), memo=memo)
 
     def __getstate__(self):
-        state = {"_hist": self._hist, "metadata": self.metadata}
-        return state
+        """
+        Version 0.8: metadata added
+        Version 0.11: version added and set to 0. metadata/_hist replaced with dict.
+
+        ``dict`` contains __dict__ without "axes" and "_hist"
+        """
+        local_dict = copy.copy(self.__dict__)
+        del local_dict["axes"]
+        # Version 0 of boost-histogram pickle state
+        return (0, local_dict)
 
     def __setstate__(self, state):
-        self._hist = state["_hist"]
-        self.metadata = state.get("metadata", None)
-        self.axes = self._generate_axes_()
+        if isinstance(state, tuple):
+            if state[0] == 0:
+                for key, value in state[1].items():
+                    self.__dict__[key] = value
+            else:
+                msg = "Cannot open boost-histogram pickle v{}".format(state[0])
+                raise RuntimeError(msg)
+
+            self.axes = self._generate_axes_()
+
+        else:  # Classic (0.10 and before) state
+            self._hist = state["_hist"]
+            self.metadata = state.get("metadata", None)
+            self.axes = self._generate_axes_()
+            for ax in self.axes:
+                ax._ax.metadata = {"metadata": ax._ax.metadata}
 
     def __repr__(self):
         newline = "\n  "