+++ src/boost_histogram/accumulators.py
@@ -11,7 +11,7 @@ for cls in (Sum, Mean, WeightedSum, WeightedMean):
     cls.__module__ = "boost_histogram.accumulators"
 del cls
 
-# Not supported by PyBind builtins
+# Not supported by pybind builtins
 # Enable if wrapper added
 # inject_signature("self, value")(Sum.fill)
 # inject_signature("self, value, *, variance=None")(WeightedSum.fill)