+++ src/register_axis.cpp
@@ -83,44 +83,30 @@ void register_axes(py::module& mod) {
                        axis::regular_uoflow_growth,
                        axis::regular_circular,
                        axis::regular_numpy>(mod, [](auto ax) {
-        ax.def(py::init<unsigned, double, double, metadata_t>(),
-               "bins"_a,
-               "start"_a,
-               "stop"_a,
-               "metadata"_a);
+        ax.def(py::init<unsigned, double, double>(), "bins"_a, "start"_a, "stop"_a);
     });
 
     register_axis<axis::regular_pow>(mod)
-        .def(py::init([](unsigned n,
-                         double start,
-                         double stop,
-                         double pow,
-                         metadata_t metadata) {
+        .def(py::init([](unsigned n, double start, double stop, double pow) {
                  return new axis::regular_pow(
-                     bh::axis::transform::pow{pow}, n, start, stop, metadata);
+                     bh::axis::transform::pow{pow}, n, start, stop);
              }),
              "bins"_a,
              "start"_a,
              "stop"_a,
-             "power"_a,
-             "metadata"_a)
+             "power"_a)
         .def_property_readonly("transform", [](const axis::regular_pow& self) {
             return self.transform();
         });
 
     register_axis<axis::regular_trans>(mod)
-        .def(py::init([](unsigned n,
-                         double start,
-                         double stop,
-                         func_transform& trans,
-                         metadata_t metadata) {
-                 return new axis::regular_trans(trans, n, start, stop, metadata);
+        .def(py::init([](unsigned n, double start, double stop, func_transform& trans) {
+                 return new axis::regular_trans(trans, n, start, stop);
              }),
              "bins"_a,
              "start"_a,
              "stop"_a,
-             "tranform"_a,
-             "metadata"_a = py::none())
+             "tranform"_a)
         .def_property_readonly("transform", [](const axis::regular_trans& self) {
             return self.transform();
         });
@@ -130,32 +116,25 @@ void register_axes(py::module& mod) {
                        axis::variable_oflow,
                        axis::variable_uoflow,
                        axis::variable_uoflow_growth,
-                       axis::variable_circular>(mod, [](auto ax) {
-        ax.def(py::init<std::vector<double>, metadata_t>(), "edges"_a, "metadata"_a);
-    });
+                       axis::variable_circular>(
+        mod, [](auto ax) { ax.def(py::init<std::vector<double>>(), "edges"_a); });
 
     register_axis_each<axis::integer_none,
                        axis::integer_uflow,
                        axis::integer_oflow,
                        axis::integer_uoflow,
                        axis::integer_growth,
-                       axis::integer_circular>(mod, [](auto ax) {
-        ax.def(py::init<int, int, metadata_t>(), "start"_a, "stop"_a, "metadata"_a);
-    });
+                       axis::integer_circular>(
+        mod, [](auto ax) { ax.def(py::init<int, int>(), "start"_a, "stop"_a); });
 
-    register_axis_each<axis::category_int, axis::category_int_growth>(mod, [](auto ax) {
-        ax.def(py::init<std::vector<int>, metadata_t>(), "categories"_a, "metadata"_a);
-    });
+    register_axis_each<axis::category_int, axis::category_int_growth>(
+        mod, [](auto ax) { ax.def(py::init<std::vector<int>>(), "categories"_a); });
 
     register_axis_each<axis::category_str, axis::category_str_growth>(mod, [](auto ax) {
-        ax.def(py::init<std::vector<std::string>, metadata_t>(),
-               "categories"_a,
-               "metadata"_a);
+        ax.def(py::init<std::vector<std::string>>(), "categories"_a);
     });
 
-    register_axis<axis::boolean>(mod, "boolean")
-        .def(py::init<>())
-        .def(py::init<metadata_t>(), "metadata"_a);
+    register_axis<axis::boolean>(mod, "boolean").def(py::init<>());
 
     ;
 }