+++ docs/CHANGELOG.md
@@ -2,10 +2,29 @@
 
 ## DEVELOP
 
+
+#### User changes
+
+* Arbitrary items can be set on an axis or histogram. [#450][]
+
+#### Bug fixes
+
 * Fixed reading pickles from boost-histogram 0.6-0.8 [#445][]
+* Minor correctness fix [#446][]
+* Accidental install of typing on Python 3.5+ fixed
+
+#### Developer changes
+* Updated to Boost 1.74 [#442][]
+* CMake installs version.py now too [#449][]
+* Updated setuptools infrastructure no longer requires NumPy [#451][]
 
 
+[#442]: https://github.com/scikit-hep/boost-histogram/pull/442
 [#445]: https://github.com/scikit-hep/boost-histogram/pull/445
+[#446]: https://github.com/scikit-hep/boost-histogram/pull/446
+[#449]: https://github.com/scikit-hep/boost-histogram/pull/449
+[#450]: https://github.com/scikit-hep/boost-histogram/pull/450
+[#451]: https://github.com/scikit-hep/boost-histogram/pull/451
 
 
 ## Version 0.10
@@ -131,7 +150,7 @@ on Boost.Histogram's C++ tools for actions like cropping.
 
 #### Developer changes
 
-* Update Boost to 1.73 [#359][], PyBind11 to 2.5.0 [#351][], Boost.Histogram to pre-1.74 [#388][]
+* Update Boost to 1.73 [#359][], pybind11 to 2.5.0 [#351][], Boost.Histogram to pre-1.74 [#388][]
 * Cropping no longer uses workaround [#373][]
 * Many more checks added to [`pre-commit`][] [#366][]
 * Deprecating `cpp` interface [#391][]
@@ -340,7 +359,7 @@ transition existing 0.5.x code to the new API.
 #### Bug fixes:
 
 * Use relative paths in setup.py to avoid resolving WSL paths on Windows [#162][], [#163][]
-* Better Pybind11 support for Python 3.8 [#168][]
+* Better pybind11 support for Python 3.8 [#168][]
 
 #### Developer changes: